% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yasibttp1.R
\name{rdist}
\alias{rdist}
\title{statistiques descriptives}
\usage{
rdist(x, p,n)
}
\arguments{
\item{x}{numeric vector representing the values of the random variable}

\item{y}{numeric vector representing the probabilities}

\item{n}{number of iterations}
}
\description{
statistiques descriptives
}
